[T11706] Port createOWLApp to TypeScript
