+++ static/src/ts/untyped.d.ts
@@ -1,0 +1,7 @@
+
+declare module 'web.core';
+declare module 'web.Dialog';
+declare module 'web.public.widget';
+declare module 'web.session';
+declare module 'web.rpc';
+declare module 'web.OwlCompatibility';
